Change project name to PressOneMillionTimes
:recycle: refactor: Change project name to PressOneMillionTimes
