+++ app/src/androidTest/java/com/merive/pressonemilliontimes/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.merive.press1mtimes;
+package com.merive.pressonemilliontimes;
 
 import android.content.Context;
