+++ app/src/main/java/com/merive/pressonemilliontimes/activities/FinishActivity.java
@@ -1,4 +1,4 @@
-package com.merive.press1mtimes.activities;
+package com.merive.pressonemilliontimes.activities;
 
 import android.content.Context;
 import android.content.Intent;
@@ -13,7 +13,7 @@ import android.widget.TextView;
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.merive.press1mtimes.R;
+import com.merive.pressonemilliontimes.R;
 
 public class FinishActivity extends AppCompatActivity {
