+++ app/src/main/java/com/merive/pressonemilliontimes/fragments/ResetFragment.java
@@ -1,25 +1,25 @@
-package com.merive.press1mtimes.fragments;
+package com.merive.pressonemilliontimes.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ImageView;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 
-import com.merive.press1mtimes.R;
-import com.merive.press1mtimes.activities.MainActivity;
+import com.merive.pressonemilliontimes.R;
+import com.merive.pressonemilliontimes.activities.MainActivity;
 
-public class SplashMessageFragment extends Fragment {
+public class ResetFragment extends Fragment {
 
-    ConstraintLayout leftButton, rightButton;
-    Button cancel;
+    TextView titleText;
+    Button cancelButton, confirmButton;
     MainActivity mainActivity;
+
 
     /**
      * Called to have the fragment instantiate its user interface view
@@ -33,12 +33,12 @@ public class SplashMessageFragment extends Fragment {
      */
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_splash_message, parent, false);
+        return inflater.inflate(R.layout.fragment_reset, parent, false);
     }
 
     /**
      * Called immediately after onCreateView has returned, but before any saved state has been restored in to the view
-     * There initializes basic variables, sets click listeners for buttons
+     * There initializes basic variables, sets button listeners
      *
      * @param view               The View returned by onCreateView
      * @param savedInstanceState If non-null, this fragment is being re-constructed from a previous saved state as given here
@@ -59,26 +59,25 @@ public class SplashMessageFragment extends Fragment {
      * @see View
      */
     private void initVariables() {
-        rightButton = getView().findViewById(R.id.splash_right_position_button);
-        leftButton = getView().findViewById(R.id.splash_left_position_button);
-        cancel = getView().findViewById(R.id.splash_message_cancel_button);
+        titleText = getView().findViewById(R.id.reset_title);
+        cancelButton = getView().findViewById(R.id.reset_cancel_button);
+        confirmButton = getView().findViewById(R.id.reset_confirm_button);
         mainActivity = ((MainActivity) getActivity());
     }
 
     /**
-     * This method sets click listeners for rightButton and leftButton
+     * Sets button click listeners
      *
-     * @see ImageView
+     * @see Button
      */
     private void setListeners() {
-        rightButton.setOnClickListener(v -> setPosition(0.98f, 0.98f));
-        leftButton.setOnClickListener(v -> setPosition(0.02f, 0.98f));
-        cancel.setOnClickListener(v -> clickCancel());
+        cancelButton.setOnClickListener(v -> clickCancel());
+        confirmButton.setOnClickListener(v -> clickConfirm());
     }
 
     /**
      * Executes when clicking on cancelButton
-     * Makes vibration effect and closes SplashMessageFragment
+     * Makes vibration effect and closes ResetFragment
      */
     private void clickCancel() {
         mainActivity.makeVibration(1);
@@ -86,15 +85,14 @@ public class SplashMessageFragment extends Fragment {
     }
 
     /**
-     * This methods saves Splash Message Position in sharedPreferences memory
-     *
-     * @param horizontal Horizontal float position value
-     * @param vertical   Vertical float position value
+     * Executes when clicking on confirmButton
+     * Makes vibration, resets counter value (sets default value (default value is 000000)) and makes toast message
      */
-    private void setPosition(float horizontal, float vertical) {
+    private void clickConfirm() {
         mainActivity.makeVibration(1);
-        mainActivity.setSplashPosition(horizontal, vertical);
-        mainActivity.makeToast(getResources().getString(R.string.splash_position_changed));
+        mainActivity.resetScore();
+        mainActivity.setScoreToCounter();
+        mainActivity.makeToast("The Score has been reset");
         mainActivity.setFragment(new SettingsFragment());
     }
 }
