+++ app/src/main/res/layout/activity_finish.xml
@@ -35,7 +35,7 @@
         android:fontFamily="@font/odibee"
         android:gravity="center"
         android:singleLine="true"
-        android:text="@string/finish"
+        android:text="@string/app_name"
         android:textAllCaps="true"
         android:textColor="@color/red"
         android:textSize="@dimen/font_large_size"
