+++ app/src/main/res/values/strings.xml
@@ -1,10 +1,9 @@
 <resources>
     <!--  MainActivity  -->
-    <string name="app_name">Press1MTimes</string>
+    <string name="app_name">PressOneMillionTimes</string>
 
     <!--  FinishActivity  -->
-    <string name="finish">Finish Press1MTimes</string>
-    <string name="congratulations">Congratulations,\nyou completed Press1MTimes!</string>
+    <string name="congratulations">Congratulations,\nyou completed PressOneMillionTimes!</string>
     <string name="hope">I hope you enjoyed it</string>
 
     <!--  SettingsFragment  -->
@@ -30,7 +29,7 @@
     <string name="scoreshare">ScoreShare</string>
     <string name="scan">Scan QR</string>
     <string name="qr_error">QR-Code cannot be scanning</string>
-    <string name="score_changed">Press1MTimes Score was changed</string>
+    <string name="score_changed">PressOneMillionTimes Score was changed</string>
 
     <!-- IconsFragment -->
     <string name="icons">Icons</string>
