+++ app/src/test/java/com/merive/pressonemilliontimes/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.merive.press1mtimes;
+package com.merive.pressonemilliontimes;
 
 import org.junit.Test;
