+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = "Press1MTimes"
+rootProject.name = "PressOneMillionTimes"
 include ':app'
